refactor: move secure storage extensions to global extensions package
