+++ lib/extensions/extensions.dart
@@ -13,6 +13,7 @@ export 'messaging_socket_exception.dart';
 export 'platform_deep_link.dart';
 export 'request_failure.dart';
 export 'route_match.dart';
+export 'secure_storage_extension.dart';
 export 'session_status.dart';
 export 'signaling_exception.dart';
 export 'signaling_hangup_failure.dart';
